api/requirements.txt
# Beach Monitor API Service Requirements

# Core API framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3

# Computer Vision
opencv-python==4.8.1.78
ultralytics==8.0.196
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0

# AWS and utilities
boto3==1.29.7
python-dotenv==1.0.0
PyYAML==6.0.1
Pillow==10.1.0
numpy==1.24.3

# YouTube capture
yt-dlp==2023.11.16
pytz==2023.3

# Optional: Rate limiting
slowapi==0.1.9

# Optional: Monitoring
prometheus-client==0.19.0
